FLICK-34 Merge Map, Contact and Feed in to one. Added pager library Updated compose version to 1.4.0-alpha03 Renamed several files and function
