+++ app/src/main/java/com/ikuzMirel/flick/ui/mainContent/map/MapScreen.kt
@@ -1,0 +1,31 @@
+package com.ikuzMirel.flick.ui.mainContent.map
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.google.android.gms.maps.model.MapStyleOptions
+import com.google.maps.android.compose.GoogleMap
+import com.google.maps.android.compose.MapProperties
+import com.google.maps.android.compose.MapUiSettings
+
+@Composable
+fun Map() {
+    GoogleMap(
+        modifier = Modifier
+            .fillMaxSize(),
+        uiSettings = MapUiSettings(
+            mapToolbarEnabled = false,
+            zoomControlsEnabled = false
+        ),
+        properties = MapProperties(
+            mapStyleOptions = MapStyleOptions(
+                when{
+                    isSystemInDarkTheme() -> MapStyle.nightMap
+                    else -> MapStyle.dayMap
+                }
+            )
+        ),
+    )
+}
+
