+++ app/src/main/java/com/ikuzMirel/flick/ui/mainContent/story/FeedScreen.kt
@@ -1,57 +1,40 @@
-package com.ikuzMirel.flick.ui.story
+package com.ikuzMirel.flick.ui.mainContent.story
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.outlined.*
-import androidx.compose.material3.*
+import androidx.compose.material.icons.outlined.ChatBubbleOutline
+import androidx.compose.material.icons.outlined.FavoriteBorder
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconToggleButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.ikuzMirel.flick.R
-import com.ikuzMirel.flick.ui.components.topAppBars.NavOnlyTopBar
 import com.ikuzMirel.flick.ui.theme.cocogooseLight
-import kotlin.random.Random
 
 // TODO: Find a better name for the screen
 // TODO: Hardcoded data source, change when viewModel is created
-@OptIn(ExperimentalMaterial3Api::class)
+
 @Composable
-fun Story() {
-    Scaffold(
-        topBar = { NavOnlyTopBar {} }, //TODO: Change when chat screen is done
-        bottomBar = { NavigationBar() {} }, //TODO: Change when chat screen is done
+fun Feed() {
+    LazyColumn(
         modifier = Modifier
             .fillMaxSize()
-    ) { paddingValues ->
-        Content(
-            modifier = Modifier
-                .padding(paddingValues)
-                .fillMaxSize()
-        )
-    }
-}
-
-@Composable
-private fun Content(
-    modifier: Modifier,
-) {
-
-    LazyColumn(
-        modifier = modifier
+            .padding(top = 64.dp)
     ){
         items(5){
-            StoryItem(
+            FeedItem(
                 user = "Xuan",
                 game = "Valorant",
                 isLikeClicked = {},
@@ -62,7 +45,7 @@ private fun Content(
 }
 
 @Composable
-fun StoryItem(
+fun FeedItem(
     user: String,
     game: String,
     isLikeClicked: () -> Unit,
