+++ app/src/main/java/com/ikuzMirel/flick/ui/theme/Font.kt
@@ -10,5 +10,5 @@ val cocogooseBold = FontFamily(Font(R.font.cocogoose, FontWeight.Bold))
 val cocogooseSemiBold = FontFamily(Font(R.font.cocogoose_semilight, FontWeight.SemiBold))
 val cocogooseRegular = FontFamily(Font(R.font.cocogoose_light, FontWeight.Normal))
 val cocogooseSemiLight = FontFamily(Font(R.font.cocogoose_semilight, FontWeight.ExtraLight))
-val cocogooseLight = FontFamily(Font(R.font.cocogoose_light, FontWeight.Thin))
+val cocogooseLight = FontFamily(Font(R.font.cocogoose_ultralight, FontWeight.Thin))
 val museoRegular = FontFamily(Font(R.font.museosans_500, FontWeight.W500))
