+++ app/src/main/java/com/ikuzMirel/flick/ui/theme/Type.kt
@@ -1,10 +1,6 @@
 package com.ikuzMirel.flick.ui.theme
 
 import androidx.compose.material3.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
 
 // Set of Material typography styles to start with
 val Typography = Typography(
