+++ versions.properties
@@ -7,21 +7,21 @@
 #### suppress inspection "SpellCheckingInspection" for whole file
 #### suppress inspection "UnusedProperty" for whole file
 
-version.androidx.compose.animation=1.4.0-alpha02
+version.androidx.compose.animation=1.4.0-alpha03
 
-version.androidx.compose.foundation=1.4.0-alpha02
+version.androidx.compose.foundation=1.4.0-alpha03
 
-version.androidx.compose.material=1.4.0-alpha02
-##                    # available=1.4.0-alpha01
-##                    # available=1.4.0-alpha02
+version.androidx.compose.material=1.4.0-alpha03
 
-version.androidx.compose.runtime=1.4.0-alpha02
+version.androidx.compose.runtime=1.4.0-alpha03
 
 version.coil-kt=2.2.2
 
-version.google.accompanist=0.27.0
+version.google.accompanist=0.28.0
 
-version.google.android.maps-compose=2.7.2
+version.google.android.maps-compose=2.7.3
+##                      # available=2.7.4
+##                      # available=2.8.0
 
 version.google.android.play-services-maps=18.1.0
 
@@ -29,7 +29,10 @@ version.google.android.play-services-maps=18.1.0
 version.org.jacoco..org.jacoco.ant=0.8.7
 
 version.kotlin=1.7.21
+## # available=1.7.22
 ## # available=1.8.0-Beta
+## # available=1.8.0-RC
+## # available=1.8.0-RC2
 
 version.junit.junit=4.13.2
 
@@ -45,8 +48,10 @@ version.androidx.lifecycle=2.5.1
 version.androidx.core=1.9.0
 
 version.androidx.compose.ui=1.4.0-alpha02
+##              # available=1.4.0-alpha03
 
 version.androidx.compose.material3=1.1.0-alpha02
+##                     # available=1.1.0-alpha03
 
 version.androidx.compose.compiler=1.4.0-alpha02
